Generate an envisioning workshop agenda
In this I want you be a Technical Program Manager that will host an Envisioning Workshop. An Envisioning workshop is a designing thinking style hybrid workshop where we have attendees both in person and remote that will consist of Initial understanding of Customer’s business goal, the engagement summary, and technical challenges, Identified stakeholders from Customer, Identified Customer or Partner engineering resources to Code With, Customer location and timeline, Value and risk markers, Tech scenario, Start of legal and/or data exploration process.  Also I want you to include a table that you will pre-populate for ideas for features and opportunities that we can co-engineer with our team and the customer to develop based on this proposed project.  This table has a table for Feasibility, which is how feasibility is this opportunity to do and the positive Impact, meaning how much Impact will this have to help the company.  You can generate at least 10 opportunities for this table to get the team thinking about this.  
 
Create an agenda that we can cover in one day, we'll use a mural board, so consider how we can best lay out the sections and collectively capture notes and share with the board. Provide details for the agenda, how much time, what roles on the project lead sections, and provide suggestions on how to make this a fun and interactive experience for everyone. 
 
The project is about <insert here your details about your project>